requirements.txt
# devops
black==22.3.0  # code formatter
click==8.1.3  # for command line interfaces

pytest==7.1.3
pytest-cov==4.0.0  # for test coverage
nbval==0.10.0  # for testing jupyter notebooks
nbqa==1.7.0  # for running black, flake8, isort, mypy, pylint on notebooks

pylint==2.15.3
ruff==0.0.290  # for running pylint on notebooks

requests==2.26.0  # for making http requests

databricks-sql-connector==3.5.0  # for connecting to databricks sql
python-dotenv==1.0.1  # for loading environment variables from .env file

# web
# fastapi==0.111.0  # used when run docs.docker demo
# markdown==3.4.1


# data
numpy==1.22.0
pandas==2.0.3
# polars==1.7.1  # advanced data manipulation than pandas
# matplotlib==3.7.1
tabulate==0.8.9  # for pretty printing dataframes
# seaborn==0.11.2  # for pretty plots
